base: nat464X: Only update ipv4 pseudo-interface when it's connected
On some devices, the network link state may go up even though there
isn't going to be an active attempt to connect to the network
due to other networks taking precedence.

Currently, when this happens the network link is forced to attempt to
connect due to notification sent to the ConnectivityService.

Rather, let's wait until there is definitely an IP address and then
notify it at that time.

This solves a bad battery drain that occurs sometimes on victara
when connected to WiFi due to the mobile data network going through
constant connect/disconnect cycles.

Change-Id: Iee390f948109b3412ec86d1b3c5fbc53a4859b7b